+++ ebl/fragmentarium/infrastructure/fragment_sign_matcher.py
@@ -1,5 +1,5 @@
 from typing import List, Dict
-from ebl.fragmentarium.infrastructure.fragment_lemma_matcher import flatten_field
+from ebl.common.query.util import flatten_field
 
 
 class SignMatcher:
